Replaced duplicated files by symlinks
